Refactor markdown-remark integration to optimize component exports and update type definition
