+++ packages/markdown-remark/src/index.ts
@@ -1,2 +1,2 @@
 export * from './processor/index.js';
-export { markdownRemark, markdownRemark as default } from './integration/index.js';
+export { markdownRemark, default, type MarkdownRemarkOptions } from './integration/index.js';
